change status of initital steps and update the step order

+++ API/src/main/java/com/beit/aquint/project/projectprocess/entity/Projects.java
@@ -39,7 +39,7 @@ public class Projects extends EntityAuditInfo implements Serializable {
     private String projectDisplayName;
 
     @Column(name = "initialStepsStatus")
-    private Long initialStepsStatus;
+    private Integer initialStepsStatus;
 
     public Projects(Long tenderId,String projectDisplayName) {
         this.projectDisplayName = projectDisplayName;
